Add AllocationFreeScope
