+++ ClrHeapAllocationsAnalyzer.Test/AllocationAnalyzerTests.cs
@@ -53,7 +53,7 @@ namespace ClrHeapAllocationAnalyzer.Test
         protected Info ProcessCode(DiagnosticAnalyzer analyzer, string sampleProgram,
             ImmutableArray<SyntaxKind> expected, bool allowBuildErrors = false, string filePath = "")
         {
-            var options = new CSharpParseOptions(kind: SourceCodeKind.Script);
+            var options = new CSharpParseOptions(kind: SourceCodeKind.Script, languageVersion: LanguageVersion.CSharp8);
             var tree = CSharpSyntaxTree.ParseText(sampleProgram, options, filePath);
             var compilation = CSharpCompilation.Create("Test", new[] { tree }, references);
