+++ ClrHeapAllocationsAnalyzer.Test/IgnoreTests.cs
@@ -42,146 +42,9 @@ namespace ClrHeapAllocationAnalyzer.Test
         }        
         
         [TestMethod]
-        public void AnalyzeProgram_NotAllowCallingOwnedUnflagedMethod()
-        {
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-                
-                public string CreateString() {
-                    return new string('a', 5);
-                }
-
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public void PerfCritical() {
-                    string str = CreateString();
-                }";
-
-            var analyser = new MethodCallAnalyzer();
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.InvocationExpression));
-            Assert.AreEqual(1, info.Allocations.Count);
-        }      
-        
-        [TestMethod]
-        public void AnalyzeProgram_MethodShouldOnlyBeAllowedToCallNonAllocatingMethods()
-        {
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-                
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public int StringLength(string str) {
-                    return str.Length;
-                }
-
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public void PerfCritical(string str) {
-                    int l = StringLength(str);
-                }";
-
-            var analyser = new MethodCallAnalyzer();
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.InvocationExpression));
-            Assert.AreEqual(0, info.Allocations.Count);
-        }
-        
-        [TestMethod]
-        public void AnalyzeProgram_MethodShouldOnlyBeAllowedToCallNonAllocatingMethodsOnNonAllocatingInterface()
-        {
-            //language=cs
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-
-                interface IFoo
-                {
-                    [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                    int StringLength(string str);
-                }
-
-                public class Foo : IFoo
-                {
-                     public int StringLength(string str)
-                     {
-                        return str.Length;
-                     }
-                }
-                
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public void PerfCritical(Foo foo, string str)
-                {
-                    int l = foo.StringLength(str);
-                }";
-
-            var analyser = new MethodCallAnalyzer();
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.InvocationExpression));
-            Assert.AreEqual(0, info.Allocations.Count);
-        }
-        
-        
-        
-        [TestMethod]
-        public void AnalyzeProgram_NotAllowCallingExternalMethod()
-        {
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-                
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public string PerfCritical(string str) {
-                    return string.Copy(str);
-                }";
-
-            var analyser = new ExplicitAllocationAnalyzer();
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.InvocationExpression));
-            Assert.AreEqual(1, info.Allocations.Count);
-        }
-        
-        [TestMethod]
-        public void AnalyzeProgram_NotAllowCallingExternalMethod_UnlessItIsWhitelisted()
-        {
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-                
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public int PerfCritical(string str) {
-                    return str.Length;
-                }";
-
-            var analyser = new ExplicitAllocationAnalyzer();
-            
-            // TODO Add to whitelist
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.ObjectInitializerExpression));
-            Assert.AreEqual(1, info.Allocations.Count);
-        }
-        
-        [TestMethod]
-        public void AnalyzeProgram_NotAllowCallingExternalMethod_UnlessItIsInSafeScope()
-        {
-            const string sample =
-                @"using System;
-                using ClrHeapAllocationAnalyzer;
-                
-                [ClrHeapAllocationAnalyzer.RestrictedAllocation]
-                public int PerfCritical(string str) {
-                    using (new AllocationFreeScope());
-                    return str.Length;
-                }";
-
-            var analyser = new ExplicitAllocationAnalyzer();
-            
-            var info = ProcessCode(analyser, sample, ImmutableArray.Create(SyntaxKind.ObjectInitializerExpression));
-            Assert.AreEqual(1, info.Allocations.Count);
-        }
-        
-        [TestMethod]
         public void AnalyzeProgram_AnalyzeImplementationWhenInterfaceHasRestrictedAttributes()
         {
+            //language=cs
             const string sample =
                 @"using System;
                 using ClrHeapAllocationAnalyzer;
@@ -208,6 +71,7 @@ namespace ClrHeapAllocationAnalyzer.Test
         [TestMethod]
         public void AnalyzeProgram_AnalyzeExplicitImplementationWhenInterfaceHasRestrictedAttributes()
         {
+            //language=cs
             const string sample =
                 @"using System;
                 using ClrHeapAllocationAnalyzer;
@@ -234,6 +98,7 @@ namespace ClrHeapAllocationAnalyzer.Test
         [TestMethod]
         public void AnalyzeProgram_AnalyzeImplementationWhenBaseHasRestrictedAttributes()
         {
+            //language=cs
             const string sample =
                 @"using System;
                 using ClrHeapAllocationAnalyzer;
@@ -260,6 +125,7 @@ namespace ClrHeapAllocationAnalyzer.Test
         [TestMethod]
         public void AnalyzeProgram_AnalyzeImplementationWhenRootHasRestrictedAttributes()
         {
+            //language=cs
             const string sample =
                 @"using System;
                 using ClrHeapAllocationAnalyzer;
@@ -291,6 +157,7 @@ namespace ClrHeapAllocationAnalyzer.Test
         [TestMethod]
         public void AnalyzeProgram_AnalyzeImplementationWhenRootInterfaceHasRestrictedAttributes()
         {
+            //language=cs
             const string sample =
                 @"using System;
                 using ClrHeapAllocationAnalyzer;
